Create a file for inference tip
This permit to not have to import wrapt in __init__.py

It's a dependencie and we want to be able to use attr: astroid.__version__ for packaging without installing depdencies first.
